Update types in os_cpu.h

+++ src/ucos-ii/ports/os_cpu.h
@@ -42,8 +42,8 @@ typedef unsigned char  INT8U;                    /* Unsigned  8 bit quantity    
 typedef signed   char  INT8S;                    /* Signed    8 bit quantity                           */
 typedef unsigned short INT16U;                   /* Unsigned 16 bit quantity                           */
 typedef signed   short INT16S;                   /* Signed   16 bit quantity                           */
-typedef unsigned int   INT32U;                   /* Unsigned 32 bit quantity                           */
-typedef signed   int   INT32S;                   /* Signed   32 bit quantity                           */
+typedef unsigned long  INT32U;                   /* Unsigned 32 bit quantity                           */
+typedef signed   long  INT32S;                   /* Signed   32 bit quantity                           */
 typedef float          FP32;                     /* Single precision floating point                    */
 typedef double         FP64;                     /* Double precision floating point                    */
